+++ src/handler/http/request/websocket/search.rs
@@ -28,7 +28,7 @@ use config::{
 use infra::errors::{Error, ErrorCodes};
 use tracing::Instrument;
 
-use super::utils::cancellation_registry_cache_utils;
+use super::{sort::order_search_results, utils::cancellation_registry_cache_utils};
 #[allow(unused_imports)]
 use crate::handler::http::request::websocket::utils::enterprise_utils;
 use crate::{
@@ -43,7 +43,7 @@ use crate::{
     },
     handler::http::request::websocket::{
         session::send_message,
-        utils::{determine_sort_column, TimeOffset, WsServerEvents},
+        utils::{TimeOffset, WsServerEvents},
     },
     service::search::{
         self as SearchService, cache, datafusion::distributed_plan::streaming_aggs_exec, sql::Sql,
@@ -470,6 +470,7 @@ async fn handle_cache_responses_and_deltas(
                     cached,
                     accumulated_results,
                     &mut curr_res_size,
+                    req.fallback_order_by_col.clone(),
                 )
                 .await?;
                 cached_resp_iter.next();
@@ -504,6 +505,7 @@ async fn handle_cache_responses_and_deltas(
                 cached,
                 accumulated_results,
                 &mut curr_res_size,
+                req.fallback_order_by_col.clone(),
             )
             .await?;
         }
@@ -606,6 +608,7 @@ async fn process_delta(
         );
 
         if !search_res.hits.is_empty() {
+            search_res = order_search_results(search_res, req.fallback_order_by_col);
             // for every partition, compute the queried range omitting the result cache ratio
             let queried_range =
                 calc_queried_range(start_time, end_time, search_res.result_cache_ratio);
@@ -745,6 +748,7 @@ async fn send_cached_responses(
     cached: &CachedQueryResponse,
     accumulated_results: &mut Vec<SearchResultType>,
     curr_res_size: &mut i64,
+    fallback_order_by_col: Option<String>,
 ) -> Result<(), Error> {
     if cancellation_registry_cache_utils::is_cancelled(trace_id) {
         log::info!(
@@ -777,6 +781,8 @@ async fn send_cached_responses(
             cached.cached_response.total = cache_hits;
         }
     }
+
+    cached.cached_response = order_search_results(cached.cached_response, fallback_order_by_col);
 
     // Accumulate the result
     accumulated_results.push(SearchResultType::Cached(cached.cached_response.clone()));
@@ -1079,149 +1085,4 @@ async fn write_results_to_cache(
     }
 
     Ok(())
-}
-
-/// Determines and applies sorting to search results
-fn order_search_results(
-    mut search_res: Response,
-    fallback_order_by_col: Option<String>,
-) -> Response {
-    if search_res.hits.is_empty() {
-        return search_res;
-    }
-
-    // First determine the strategy
-    let strategy = determine_sort_strategy(&search_res, fallback_order_by_col);
-
-    // Then apply it
-    apply_sort_strategy(&mut search_res, strategy);
-
-    search_res
-}
-
-/// Represents different sorting strategies
-enum SortStrategy {
-    SqlOrderBy,
-    FallbackColumn(String, OrderBy),
-    AutoDetermine(String, bool), // (column, is_string)
-    NoSort,
-}
-
-/// Determines which sorting strategy to use
-fn determine_sort_strategy(
-    search_res: &Response,
-    fallback_order_by_col: Option<String>,
-) -> SortStrategy {
-    // Check SQL ORDER BY first
-    if let Some(order_by) = search_res.order_by {
-        log::info!(
-            "[trace_id: {}] Using user-specified ORDER BY: {:?}",
-            &search_res.trace_id,
-            order_by
-        );
-        return SortStrategy::SqlOrderBy;
-    }
-
-    // Check fallback column
-    if let Some(col) = find_fallback_column(search_res, fallback_order_by_col) {
-        return SortStrategy::FallbackColumn(col, OrderBy::Desc);
-    }
-
-    // Auto-determine as last resort
-    if let Some((col, is_string)) = determine_sort_column(&search_res.hits[0]) {
-        log::info!(
-            "[trace_id: {}] Auto-sorting by column: {}, type: {}",
-            &search_res.trace_id,
-            col,
-            if is_string { "string" } else { "numeric" }
-        );
-        return SortStrategy::AutoDetermine(col, is_string);
-    }
-
-    SortStrategy::NoSort
-}
-
-/// Finds and validates fallback column in results
-fn find_fallback_column(search_res: &Response, fallback_col: Option<String>) -> Option<String> {
-    let fallback_col = fallback_col?;
-    let first_hit = search_res.hits.first()?.as_object()?;
-
-    // Find case-insensitive match
-    first_hit
-        .keys()
-        .find(|k| k.eq_ignore_ascii_case(&fallback_col))
-        .map(|k| {
-            log::info!(
-                "[trace_id: {}] Using fallback ORDER BY: {}",
-                &search_res.trace_id,
-                k
-            );
-            k.to_string()
-        })
-}
-
-/// Applies the chosen sort strategy to results
-fn apply_sort_strategy(search_res: &mut Response, strategy: SortStrategy) {
-    match strategy {
-        SortStrategy::SqlOrderBy => (), // Already sorted
-        SortStrategy::FallbackColumn(col, order) => {
-            sort_by_column(search_res, &col, true, order == OrderBy::Desc);
-            if search_res.order_by.is_none() {
-                search_res.order_by = Some(order);
-            }
-        }
-        // SortStrategy::AutoDetermine(col, is_string) => {
-        //     sort_by_column(search_res, &col, is_string, !is_string);
-        // }
-        SortStrategy::NoSort => (),
-        _ => (),
-    }
-}
-
-/// Sorts results by a specific column
-fn sort_by_column(search_res: &mut Response, column: &str, is_string: bool, descending: bool) {
-    search_res.hits.sort_by(|a, b| {
-        let ordering = if is_string {
-            compare_string_values(a, b, column)
-        } else {
-            compare_numeric_values(a, b, column)
-        };
-        if descending {
-            ordering.reverse()
-        } else {
-            ordering
-        }
-    });
-}
-
-/// Compares string values with null handling
-fn compare_string_values(
-    a: &serde_json::Value,
-    b: &serde_json::Value,
-    column: &str,
-) -> std::cmp::Ordering {
-    let a_val = a.get(column).and_then(|v| v.as_str());
-    let b_val = b.get(column).and_then(|v| v.as_str());
-    match (a_val, b_val) {
-        (Some(a), Some(b)) => a.cmp(b),
-        (None, Some(_)) => std::cmp::Ordering::Greater,
-        (Some(_), None) => std::cmp::Ordering::Less,
-        (None, None) => std::cmp::Ordering::Equal,
-    }
-}
-
-/// Compares numeric values with null handling
-fn compare_numeric_values(
-    a: &serde_json::Value,
-    b: &serde_json::Value,
-    column: &str,
-) -> std::cmp::Ordering {
-    let a_val = a.get(column).and_then(|v| v.as_f64());
-    let b_val = b.get(column).and_then(|v| v.as_f64());
-    match (a_val, b_val) {
-        (Some(a), Some(b)) => a.partial_cmp(&b).unwrap_or(std::cmp::Ordering::Equal),
-        (None, Some(_)) => std::cmp::Ordering::Greater,
-        (Some(_), None) => std::cmp::Ordering::Less,
-        (None, None) => std::cmp::Ordering::Equal,
-    }
 }
